+++ packages/drizzle/test/resolver-postgres.spec.ts
@@ -90,22 +90,12 @@ describe("resolver by postgres", () => {
   describe.concurrent("query", () => {
     it("should query users correctly", async () => {
       const q = /* GraphQL */ `
-        query users(
-          $orderBy: UsersOrderBy
-          $where: UsersFilters!
-          $limit: Int
-          $offset: Int
-        ) {
-          users(
-            orderBy: $orderBy
-            where: $where
-            limit: $limit
-            offset: $offset
-          ) {
-            id
-            name
-          }
+      query users ($orderBy: UserOrderBy!, $where: UserFilters!, $limit: Int, $offset: Int) {
+        users(orderBy: $orderBy, where: $where, limit: $limit, offset: $offset) {
+          id
+          name
         }
+      }
       `
       await expect(
         execute(q, {
@@ -142,16 +132,12 @@ describe("resolver by postgres", () => {
       await expect(
         execute(
           /* GraphQL */ `
-            query users(
-              $orderBy: UsersOrderBy
-              $where: UsersFilters!
-              $offset: Int
-            ) {
-              usersSingle(orderBy: $orderBy, where: $where, offset: $offset) {
-                id
-                name
-              }
+          query users ($orderBy: UserOrderBy, $where: UserFilters!, $offset: Int) {
+            usersSingle(orderBy: $orderBy, where: $where, offset: $offset) {
+              id
+              name
             }
+          }
           `,
           {
             where: { name: { eq: "Taylor" } },
@@ -164,18 +150,8 @@ describe("resolver by postgres", () => {
 
     it("should query user with posts correctly", async () => {
       const q = /* GraphQL */ `
-        query users(
-          $orderBy: UsersOrderBy
-          $where: UsersFilters!
-          $limit: Int
-          $offset: Int
-        ) {
-          users(
-            orderBy: $orderBy
-            where: $where
-            limit: $limit
-            offset: $offset
-          ) {
+        query users ($orderBy: UserOrderBy, $where: UserFilters!, $limit: Int, $offset: Int) {
+          users(orderBy: $orderBy,where: $where, limit: $limit, offset: $offset) {
             id
             name
             posts {
@@ -213,7 +189,7 @@ describe("resolver by postgres", () => {
   describe("mutation", () => {
     it("should insert a new user correctly", async () => {
       const q = /* GraphQL */ `
-        mutation insertIntoUsers($values: [UsersInsertInput!]!) {
+        mutation insertIntoUsers($values: [UserInsertInput!]!) {
           insertIntoUsers(values: $values) {
             id
             name
@@ -236,12 +212,107 @@ describe("resolver by postgres", () => {
       expect(Tina).toBeDefined()
     })
 
+    it("should insert a user with on conflict correctly", async () => {
+      const q = /* GraphQL */ `
+        mutation insertIntoUsers($values: [UserInsertInput!]!, $doNothing: UserInsertOnConflictDoNothingInput, $doUpdate: UserInsertOnConflictDoUpdateInput) {
+          insertIntoUsers(onConflictDoNothing: $doNothing, onConflictDoUpdate: $doUpdate, values: $values) {
+            id
+            name
+          }
+        }
+      `
+
+      await expect(
+        execute(q, {
+          values: [{ name: "Tina", id: 77 }],
+        })
+      ).resolves.toMatchObject({
+        insertIntoUsers: [{ name: "Tina" }],
+      })
+
+      await expect(
+        execute(q, {
+          values: [{ name: "Tina", id: 77 }],
+          doNothing: {},
+        })
+      ).resolves.toMatchObject({
+        insertIntoUsers: [],
+      })
+
+      await expect(
+        execute(q, {
+          values: [{ name: "Tina", id: 77 }],
+          doNothing: { target: ["id"] },
+        })
+      ).resolves.toMatchObject({
+        insertIntoUsers: [],
+      })
+
+      await expect(
+        execute(q, {
+          values: [{ name: "TinaInsert", id: 77 }],
+          doUpdate: {
+            target: ["id"],
+            set: { name: "TinaUpdate" },
+          },
+        })
+      ).resolves.toMatchObject({
+        insertIntoUsers: [{ name: "TinaUpdate" }],
+      })
+    })
+
+    it("should insert a single user with on conflict correctly", async () => {
+      const q = /* GraphQL */ `
+        mutation insertIntoUsersSingle($value: UserInsertInput!, $doNothing: UserInsertOnConflictDoNothingInput, $doUpdate: UserInsertOnConflictDoUpdateInput) {
+          insertIntoUsersSingle(onConflictDoNothing: $doNothing, onConflictDoUpdate: $doUpdate, value: $value) {
+            id
+            name
+          }
+        }
+      `
+
+      await expect(
+        execute(q, {
+          value: { name: "Tina", id: 78 },
+        })
+      ).resolves.toMatchObject({
+        insertIntoUsersSingle: { name: "Tina" },
+      })
+
+      await expect(
+        execute(q, {
+          value: { name: "Tina", id: 78 },
+          doNothing: {},
+        })
+      ).resolves.toMatchObject({
+        insertIntoUsersSingle: null,
+      })
+
+      await expect(
+        execute(q, {
+          value: { name: "Tina", id: 78 },
+          doNothing: { target: ["id"] },
+        })
+      ).resolves.toMatchObject({
+        insertIntoUsersSingle: null,
+      })
+
+      await expect(
+        execute(q, {
+          value: { name: "Tina", id: 78 },
+          doUpdate: {
+            target: ["id"],
+            set: { name: "TinaUpdate" },
+          },
+        })
+      ).resolves.toMatchObject({
+        insertIntoUsersSingle: { name: "TinaUpdate" },
+      })
+    })
+
     it("should update user information correctly", async () => {
       const q = /* GraphQL */ `
-        mutation updateUsers(
-          $set: UsersUpdateInput!
-          $where: UsersFilters!
-        ) {
+        mutation updateUsers($set: UserUpdateInput!, $where: UserFilters!) {
           updateUsers(set: $set, where: $where) {
             id
             name
@@ -276,7 +347,7 @@ describe("resolver by postgres", () => {
 
     it("should delete a user correctly", async () => {
       const q = /* GraphQL */ `
-        mutation deleteFromUsers($where: UsersFilters!) {
+        mutation deleteFromUsers($where: UserFilters!) {
           deleteFromUsers(where: $where) {
             id
             name
@@ -306,7 +377,7 @@ describe("resolver by postgres", () => {
 
     it("should insert a new post correctly", async () => {
       const q = /* GraphQL */ `
-        mutation insertIntoPosts($values: [PostsInsertInput!]!) {
+        mutation insertIntoPosts($values: [PostInsertInput!]!) {
           insertIntoPosts(values: $values) {
             id
             title
@@ -337,10 +408,7 @@ describe("resolver by postgres", () => {
 
     it("should update post information correctly", async () => {
       const q = /* GraphQL */ `
-        mutation updatePosts(
-          $set: PostsUpdateInput!
-          $where: PostsFilters!
-        ) {
+        mutation updatePosts($set: PostUpdateInput!, $where: PostFilters!) {
           updatePosts(set: $set, where: $where) {
             id
             title
@@ -376,7 +444,7 @@ describe("resolver by postgres", () => {
 
     it("should delete a post correctly", async () => {
       const q = /* GraphQL */ `
-        mutation deleteFromPosts($where: PostsFilters!) {
+        mutation deleteFromPosts($where: PostFilters!) {
           deleteFromPosts(where: $where) {
             id
             title