+++ packages/drizzle/test/schema/postgres.ts
@@ -7,7 +7,14 @@ export const users = drizzleSilk(
     name: t.text().notNull(),
     age: t.integer(),
     email: t.text(),
-  })
+  }),
+  {
+    name: "User",
+    description: "A user",
+    fields: {
+      name: { description: "The name of the user" },
+    },
+  }
 )
 
 export const posts = drizzleSilk(
@@ -16,5 +23,12 @@ export const posts = drizzleSilk(
     title: t.text().notNull(),
     content: t.text(),
     authorId: t.integer().references(() => users.id, { onDelete: "cascade" }),
-  })
+  }),
+  {
+    name: "Post",
+    description: "A post",
+    fields: {
+      title: { description: "The title of the post" },
+    },
+  }
 )